refactor: move loadLanguage to utils
